pyproject.toml
[project]
name = "wiki"
version = "0.1.0"
description = "Personal knowledge wiki web frontend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "itsdangerous>=2.2",
    "python-frontmatter>=1.1",
    "markdown>=3.7",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
